+++ mulgomanager/frontend/static/images/recordbutton.png
